:gear: Déplace reduceSteps
